+++ juniper-eager-loading/examples/has_many_through.rs
@@ -101,7 +101,11 @@ mod models {
         type Error = Box<dyn std::error::Error>;
         type Connection = DbConnection;
 
-        fn load(employments: &[i32], _: &(), db: &Self::Connection) -> Result<Vec<Self>, Self::Error> {
+        fn load(
+            employments: &[i32],
+            _: &(),
+            db: &Self::Connection,
+        ) -> Result<Vec<Self>, Self::Error> {
             unimplemented!()
         }
     }
@@ -110,7 +114,11 @@ mod models {
         type Error = Box<dyn std::error::Error>;
         type Connection = DbConnection;
 
-        fn load(employments: &[i32], _: &(), db: &Self::Connection) -> Result<Vec<Self>, Self::Error> {
+        fn load(
+            employments: &[i32],
+            _: &(),
+            db: &Self::Connection,
+        ) -> Result<Vec<Self>, Self::Error> {
             unimplemented!()
         }
     }
@@ -119,7 +127,11 @@ mod models {
         type Error = Box<dyn std::error::Error>;
         type Connection = DbConnection;
 
-        fn load(employments: &[User], _: &(), db: &Self::Connection) -> Result<Vec<Self>, Self::Error> {
+        fn load(
+            employments: &[User],
+            _: &(),
+            db: &Self::Connection,
+        ) -> Result<Vec<Self>, Self::Error> {
             unimplemented!()
         }
     }
@@ -128,7 +140,11 @@ mod models {
         type Error = Box<dyn std::error::Error>;
         type Connection = DbConnection;
 
-        fn load(employments: &[Company], _: &(), db: &Self::Connection) -> Result<Vec<Self>, Self::Error> {
+        fn load(
+            employments: &[Company],
+            _: &(),
+            db: &Self::Connection,
+        ) -> Result<Vec<Self>, Self::Error> {
             unimplemented!()
         }
     }