+++ CHANGELOG.md
@@ -8,10 +8,21 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 - Move `impl_load_from_for_diesel_{pg|mysql|sqlite}!` to proc-macros. Are fully backwards compatible but will give better errors.
 - Tweak docs for `impl_load_from_for_diesel_{pg|mysql|sqlite}!`.
+- `Association` trait has been added to abstraction over `HasOne`, `OptionHasOne`, `HasMany`, and `HasManyThrough` associations.
 
 ### Breaking changes
 
-None.
+- `EagerLoadChildrenOfType::child_ids` has been removed. Use `EagerLoadChildrenOfType::load_children` instead. See [#27](https://github.com/davidpdrsn/juniper-eager-loading/issues/27) for more context.
+- `EagerLoadChildrenOfType::ChildId` has been removed. It was only used by `child_ids` and was therefore no longer necessary.
+- `LoadResult` has been renamed to `LoadChildrenOutput`. Including `Result` in the name made it seem like it might related to errors, which it wasn't.
+    - `LoadResult::Ids` has been renamed to `LoadChildrenOutput::ChildModel` to match the changes to `EagerLoadChildrenOfType::load_children`.
+    - `LoadResult::Models` has been renamed to `LoadChildrenOutput::ChildAndJoinModels` for the same reason.
+    - The second type parameter (used for the join model) now defaults to `()`.
+- The signature of `EagerLoadChildrenOfType::is_child_of` has been changed to `parent: &Self, child: &Child, join_model: &JoinModel`. Manually pulling things out of the tuple was tedious.
+- `EagerLoadChildrenOfType::association` has been added. This methods allows for some boilerplate to be removed from `EagerLoadChildrenOfType`.
+- `EagerLoadChildrenOfType::loaded_child` and `EagerLoadChildrenOfType::assert_loaded_otherwise_failed` has been removed and implemented generically using the new `Association` trait.
+
+If you're using the derive macros for everything in your app you shouldn't have to care about any of these changes. The generated code will automatically handle them.
 
 ## [0.3.1] - 2019-10-09